requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
crispy==0.6.3
cryptography==3.4.7
cycler==0.10.0
defusedxml==0.7.1
Django==3.2.4
django-allauth==0.44.0
django-crispy-forms==1.12.0
fabio==0.11.0
gunicorn==20.1.0
h5py==3.2.1
idna==2.10
kiwisolver==1.3.1
lxml==4.6.3
matplotlib==3.4.2
numpy==1.20.3
oauthlib==3.1.1
Pillow==8.2.0
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
selenium==3.141.0
silx==0.15.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.5
whitenoise==5.2.0
